==> 01 - Data Preparation.ipynb <==
%matplotlib inline
import pandas as pd
from pandas import set_option
#Limit dataframe exibition to 8 rows
set_option("display.max_rows", 5)
import scipy as sp
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np

#### Os arquivos de entrada foram convertidos para csv para diminuir seu peso e facilitar a manipulação. O conteúdo de nenhuma variável dos arquivos foi alterado

##### Os arquivos de entrada tem o mesmo formato, então eu criei uma função que carrega cada um dos arquivos e salva em uma forma melhor para sua manipulação

A função abaixo carregará o arquivo csv, depois criará uma coluna de ano para a informação de ano que está no nome das colunas. Depois disso transformarei a coluna "vl_Arrec" para numérica. Por último salvarei o dataframe processado em um csv

def processa_arquivo(path_to_load, id_vars, var_name, value_name, value_vars, column_to_convert, path_to_save):
    # Carrega o arquivo na memória
    df = pd.read_csv(path_to_load, encoding='latin1', sep=';')
    #Como temos a informação do ano no nome das colunas, vou usar o método melt para transformar essa informação em dado e facilitar a análise
    df = pd.melt(df, id_vars=id_vars, var_name=var_name, value_name=value_name, value_vars = value_vars)
    #Vou converter a variável 'vl_arrec' para int64 para facilitar nos cálculos
    df[column_to_convert] = df[column_to_convert].fillna('0').str.replace('.','').astype(np.int64)
    # Salva o dataframe
    df.to_csv(path_to_save, encoding='latin1')

#### Processa arquivo de arrecadação por Estado

path_to_load = './src/arrec_resumido.csv'
id_vars=['cd_ibge','estado']
var_name='ano'
value_name='vl_arrec'
value_vars=['2000', '2001','2002','2003','2004','2005'\
            ,'2006','2007','2008','2009','2010','2011','2012'\
            ,'2013','2014','2015','2016']
column_to_convert = 'vl_arrec'
path_to_save = './processed/arrec_por_estado.csv'
processa_arquivo(path_to_load, id_vars, var_name, value_name, value_vars, column_to_convert, path_to_save)

#### Processa arquivo de arrecadação por Município

path_to_load = './src/arrec_por_municipio.csv'
id_vars=['cd_ibge','municipio','estado']
var_name='ano'
value_name='vl_arrec'
value_vars=['2000', '2001','2002','2003','2004','2005'\
            ,'2006','2007','2008','2009','2010','2011','2012'\
            ,'2013','2014','2015','2016']
column_to_convert = 'vl_arrec'
path_to_save = './processed/arrec_por_municipio.csv'
processa_arquivo(path_to_load, id_vars, var_name, value_name, value_vars, column_to_convert, path_to_save)

#### Processa arquivo de emissões por Estado

path_to_load = './src/vl_emitido_por_estado.csv'
id_vars=['cd_ibge','estado']
var_name='ano'
value_name='vl_arrec'
value_vars=['2000', '2001','2002','2003','2004','2005'\
            ,'2006','2007','2008','2009','2010','2011','2012'\
            ,'2013','2014','2015','2016']
column_to_convert = 'vl_arrec'
path_to_save = './processed/emissoes_por_estado.csv'
processa_arquivo(path_to_load, id_vars, var_name, value_name, value_vars, column_to_convert, path_to_save)

#### Processa arquivo de emissões por Municipio

path_to_load = './src/vl_emitido_por_municipio.csv'
id_vars=['cd_ibge','municipio','estado']
var_name='ano'
value_name='vl_arrec'
value_vars=['2000', '2001','2002','2003','2004','2005'\
            ,'2006','2007','2008','2009','2010','2011','2012'\
            ,'2013','2014','2015','2016']
column_to_convert = 'vl_arrec'
path_to_save = './processed/emissoes_por_municipio.csv'
processa_arquivo(path_to_load, id_vars, var_name, value_name, value_vars, column_to_convert, path_to_save)

#### Processa arquivo de arrecadações ignoradas

path_to_load = './src/arrec_ignorado.csv'
id_vars=['cd_ibge','estado']
var_name='ano'
value_name='vl_arrec'
value_vars=['2000', '2001','2002','2003','2004','2005'\
            ,'2006','2007','2008','2009','2010','2011','2012'\
            ,'2013','2014','2015','2016']
column_to_convert = 'vl_arrec'
path_to_save = './processed/arrec_ignorado.csv'
processa_arquivo(path_to_load, id_vars, var_name, value_name, value_vars, column_to_convert, path_to_save)
