==> 02 - Data Analysis and join.ipynb <==
%matplotlib inline
import pandas as pd
from pandas import set_option
#Limit dataframe exibition to 8 rows
set_option("display.max_rows", 5)
import scipy as sp
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np
import gc

arrec_estado = pd.read_csv('./processed/arrec_por_estado.csv', encoding='latin1', sep=',')

arrec_municipio = pd.read_csv('./processed/arrec_por_municipio.csv', encoding='latin1', sep=',')

arrec_estado

arrec_municipio.drop('Unnamed: 0', axis=1, inplace=True)

arrec_estado.drop('Unnamed: 0', axis=1, inplace=True)

##### Parece que se agrupar os dados de munícipio por estado eu terei o mesmo resultado do dataframe por estado então vou validar isso

validation = pd.DataFrame(arrec_municipio.groupby(by=['estado','ano'])['vl_arrec'].sum()).reset_index()

arrec_estado.drop('cd_ibge', axis=1, inplace=True)

arrec_estado.sort_values(['estado', 'ano'], inplace=True)

arrec_estado['estado'][1]

Temos um espaço em branco no estado do df arrec_estado

arrec_estado['estado'] = arrec_estado['estado'].str.strip()

len(arrec_estado)

len(validation)

Ambos dataframes tem o mesmo número de linhas

resultado = pd.merge(arrec_estado, validation, how='inner', left_on=['estado','ano'], right_on=['estado','ano'])

len(resultado)

O inner join retornou o mesmo número de linhas então tivemos join em todas as linhas.

resultado.columns

len(resultado.ix[resultado['vl_arrec_x']==resultado['vl_arrec_y']])

##### Como vemos acima, todas as linhas tem o mesmo valor de arrecadação, então não precisamos usar o dataframe  arrec_estado

Vamos liberar memória apagando os dataframes desnecessários

del arrec_estado
del validation
del resultado
gc.collect()

emissoes_municipio = pd.read_csv('./processed/emissoes_por_municipio.csv', encoding='latin1', sep=',')

emissoes_municipio.drop(['cd_ibge','Unnamed: 0'], axis=1, inplace=True)

len(emissoes_municipio)

len(arrec_municipio)

##### Vou juntar os valores arrecadados e os valores emitidos 

len(pd.merge(emissoes_municipio, arrec_municipio, how='inner', left_on=['municipio','estado','ano'], right_on=['municipio','estado','ano']))

Como vemos acima o inner join acontece em todas as linhas então podemos seguir em frente com ele

prev = pd.merge(emissoes_municipio, arrec_municipio, how='inner', left_on=['municipio','estado','ano'], right_on=['municipio','estado','ano'])

prev = prev.rename(columns={'vl_arrec_x': 'vl_emissao', 'vl_arrec_y': 'vl_arrecadado'})

Salvando o arquivo para futuras utilizações

prev.to_csv('./processed/prev_com_cd_ibge.csv', encoding='latin1')

prev.drop('cd_ibge', axis=1, inplace=True)

prev.to_csv('./processed/prev.csv', encoding='latin1')
