==> 06 - Prophet Part II.ipynb <==
%matplotlib inline
import pandas as pd
from pandas import set_option
#Limit dataframe exibition to 8 rows
set_option("display.max_rows", 5)
import scipy as sp
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np
import gc
from fbprophet import Prophet

df = pd.read_csv('./processed/prev_com_cd_ibge.csv', encoding='latin1')
df.drop('Unnamed: 0', axis=1, inplace=True)

df['ds'] = ''
df['ds'] = df['ano'].apply(lambda x: str(x)+'-12-31')

df['ds']=pd.to_datetime(df['ds'])

df_emissao = df.loc[:,['ds','cd_ibge', 'estado', 'municipio', 'ano','vl_emissao']]
df_emissao = df_emissao.rename(columns={'vl_emissao':'y'})

ibge = df['cd_ibge'].unique()

#valores não podem ser otimizados
index = np.argwhere(ibge==1500602)
ibge = np.delete(ibge, index)

# Arrecadação == 0 em todos os anos
index = np.argwhere(ibge==2111631)
ibge = np.delete(ibge, index)

predictions_emissoes = {}
for codigo in ibge:
    m = Prophet(weekly_seasonality=False)
    m.fit(df_emissao.loc[df_emissao['cd_ibge']==codigo, ['ds','y']])
    future = m.make_future_dataframe(periods=1, freq='A')
    forecast = m.predict(future)
    predictions_emissoes[codigo] = forecast['yhat'].tail(1)

codigo

np.argwhere(ibge==2111631)

df_emissao.loc[df_emissao['cd_ibge']==2201309]

h = Prophet(weekly_seasonality=False,yearly_seasonality=True)
h.fit(df_emissao.loc[df_emissao['cd_ibge']==2201309, ['ds','y']])

set_option("display.max_rows", 18)

df_emissao.ix[df_emissao['cd_ibge']==2111631, ['ds','y']]

codigo

df_arrecadado = df.ix[:,['ds','cd_ibge', 'estado', 'municipio', 'ano','vl_arrecadado']]
df_arrecadado = df_emissao.rename(columns={'vl_arrecadadoo':'y'})

predictions_arrecadado = {}
for codigo in ibge:
    m = Prophet()
    m.fit(df_arrecadado.ix[df_arrecadado['cd_ibge']==codigo, ['ds','y']])
    future = m.make_future_dataframe(periods=1, freq='A')
    forecast = m.predict(future)
    predictions_arrecadado[codigo] = forecast['yhat'].tail(1).values[0]

predictions_arrecadado

len(predictions_arrecadado)

len(ibge)

df_emissao.ix[df_emissao['cd_ibge']==1100031, ['ds','y']].tail(5)

m = Prophet()

m.fit(df_emissao)

future = m.make_future_dataframe(periods=1, freq='A')

forecast = m.predict(future)

forecast.columns

forecast

m.plot(forecast)

df_emissao.shape
