==> 07 - Prophet Part III - Filtered information.ipynb <==
%matplotlib inline
import pandas as pd
from pandas import set_option
#Limit dataframe exibition to 8 rows
set_option("display.max_rows", 5)
import scipy as sp
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np
import gc
from fbprophet import Prophet

df = pd.read_csv('./processed/prev_com_cd_ibge.csv', encoding='latin1')
df.drop('Unnamed: 0', axis=1, inplace=True)

df['ds'] = ''
df['ds'] = df['ano'].apply(lambda x: str(x)+'-12-31')

df['dif'] =  df['vl_arrecadado'] - df['vl_emissao']

ibge = df['cd_ibge'].unique()

df = df.rename(columns={'dif':'y'})

worst10 = df.loc[(df['ano']==2016) ,['y','cd_ibge']].sort_values(by='y', \
                                                                       ascending=True).head(10)['cd_ibge'].values

df_prophet = df.loc[df['cd_ibge'].isin(worst10),['ds','cd_ibge', 'estado', 'municipio', 'ano','y']]

predictions_worst = {}
for codigo in worst10:
    m = Prophet(weekly_seasonality=False)
    m.fit(df_prophet.loc[df_prophet['cd_ibge']==codigo, ['ds','y']])
    future = m.make_future_dataframe(periods=1, freq='A')
    forecast = m.predict(future)
    predictions_worst[codigo] = forecast['yhat'].tail(1).values[0]

predictions_worst

top10 = df.loc[(df['ano']==2016) ,['y','cd_ibge']].sort_values(by='y', \
                                                                       ascending=False).head(10)['cd_ibge'].values

df_prophet = df.loc[df['cd_ibge'].isin(top10),['ds','cd_ibge', 'estado', 'municipio', 'ano','y']]

codigo

predictions_best = {}
for codigo in top10:
    m = Prophet(weekly_seasonality=False, yearly_seasonality=True, growth='logistic')
    m.fit(df_prophet.loc[df_prophet['cd_ibge']==codigo, ['ds','y']])
    future = m.make_future_dataframe(periods=1, freq='A')
    forecast = m.predict(future)
    predictions_worst[codigo] = forecast['yhat'].tail(1).values[0]

df.loc[df['cd_ibge']==3304557]

top10 = df.loc[df['ano']==2016,['vl_emissao','cd_ibge']].sort_values(by='vl_emissao', \
                                                                     ascending=False).head(10)['cd_ibge'].values

worst10 = df.loc[(df['ano']==2016) ,['vl_emissao','cd_ibge']].sort_values(by='vl_emissao', \
                                                                       ascending=True).head(10)['cd_ibge'].values

df_emissao = df.loc[df['cd_ibge'].isin(top10),['ds','cd_ibge', 'estado', 'municipio', 'ano','vl_emissao']]
df_emissao = df_emissao.rename(columns={'vl_emissao':'y'})

top_emissoes = {}
for codigo in top10:
    m = Prophet(weekly_seasonality=False)
    m.fit(df_emissao.loc[df_emissao['cd_ibge']==codigo, ['ds','y']])
    future = m.make_future_dataframe(periods=1, freq='A')
    forecast = m.predict(future)
    predictions_emissoes[codigo] = forecast['yhat'].tail(1).values[0]

df_emissao = df.loc[df['cd_ibge'].isin(worst10),['ds','cd_ibge', 'estado', 'municipio', 'ano','vl_emissao']]
df_emissao = df_emissao.rename(columns={'vl_emissao':'y'})

worst_emissoes = {}
for codigo in worst10:
    m = Prophet(weekly_seasonality=False)
    m.fit(df_emissao.loc[df_emissao['cd_ibge']==codigo, ['ds','y']])
    future = m.make_future_dataframe(periods=1, freq='A')
    forecast = m.predict(future)
    predictions_emissoes[codigo] = forecast['yhat'].tail(1).values[0]
