==> 07 - RandonForestRegressor.ipynb <==
%matplotlib inline
import pandas as pd
from pandas import set_option
#Limit dataframe exibition to 8 rows
set_option("display.max_rows", 5)
import scipy as sp
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np
import gc
from sklearn import cross_validation as cv
from sklearn.ensemble import RandomForestRegressor

df = pd.read_csv('./processed/prev.csv', encoding='latin1')
df.drop('Unnamed: 0', axis=1, inplace=True)

df_emissao = df.ix[:,['ano', 'estado', 'municipio', 'ano','vl_emissao']]

target = 'vl_emissao'

train, test = cv.train_test_split(df_emissao, train_size = 0.8)

y_test=test['vl_emissao']
test.drop(['vl_emissao'], 1, inplace=True)

dummie = pd.get_dummies(train, prefix=['est', 'mun'], columns=['estado', 'municipio'])

predictors = [x for x in dummie.columns if x not in [target]]

rtr = RandomForestRegressor(n_estimators=10, n_jobs=-1)

rtr.fit(dummie[predictors], dummie[target])
