==> 08 - Regression.ipynb <==
%matplotlib inline
import pandas as pd
from pandas import set_option
#Limit dataframe exibition to 5 rows
set_option("display.max_rows", 10)
import scipy as sp
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np
import gc
from sklearn import datasets, linear_model

df = pd.read_csv('./processed/prev_com_cd_ibge.csv', encoding='latin1')
df.drop('Unnamed: 0', axis=1, inplace=True)

ibge = df['cd_ibge'].unique()

df2 = df

municipios = pd.DataFrame(df.groupby(by=['cd_ibge','estado','municipio']).first()).reset_index()

municipios.drop(['ano','vl_emissao','vl_arrecadado'], axis=1, inplace=True)

df2['dif'] = df2['vl_arrecadado'] - df2['vl_emissao']

Vamos predizer as emissões utilizando regressão linear

predictions_dif = {}
for codigo in ibge:
    regr = linear_model.LinearRegression(n_jobs=-1)
    X = df2.loc[df['cd_ibge']==codigo, ['ano']]
    y = df2.loc[df['cd_ibge']==codigo, ['dif']]
    regr.fit(X,y)    
    predictions_dif[codigo] = regr.predict(2017)[0][0]

df_dif = pd.DataFrame(columns = ['cd_ibge','dif'])
df_dif['cd_ibge'] = predictions_dif.keys()
df_dif['dif'] = predictions_dif.values()

predicoes_dif = pd.merge(df_dif, municipios, how='inner', on='cd_ibge')

predicoes_dif.sort_values(by='dif', ascending=False).head(10)

predicoes_dif.sort_values(by='dif', ascending=True).head(10)

predicoes_dif['dif'].sum()

regioes = pd.read_csv('./src/regioes-br.csv', encoding='latin1', sep=';')

prd = pd.merge(predicoes_dif, regioes, how='inner', on='estado')

pd.DataFrame(prd.groupby(by='regiao')['dif'].sum()).reset_index().sort_values(by='dif', ascending=False)

predicoes_dif.to_csv('./processed/predicoes_2017.csv', encoding='latin1')

predictions_emissoes = {}
for codigo in ibge:
    regr = linear_model.LinearRegression(n_jobs=-1)
    X = df.loc[df['cd_ibge']==codigo, ['ano']]
    y = df.loc[df['cd_ibge']==codigo, ['vl_emissao']]
    regr.fit(X,y)    
    predictions_emissoes[codigo] = regr.predict(2017)[0][0]

Agora vamos  predizer as arrecadações

predictions_arrecadado = {}
for codigo in ibge:
    regr = linear_model.LinearRegression(n_jobs=-1)
    X = df.loc[df['cd_ibge']==codigo, ['ano']]
    y = df.loc[df['cd_ibge']==codigo, ['vl_arrecadado']]
    regr.fit(X,y)    
    predictions_arrecadado[codigo] = regr.predict(2017)[0][0]

Agora vamos montar dataframes para facilitar a manipulação dos dados

df_arrec = pd.DataFrame(columns = ['cd_ibge','vl_arrecadado'])
df_arrec['cd_ibge'] = predictions_arrecadado.keys()
df_arrec['vl_arrecadado'] = predictions_arrecadado.values()

df_emis = pd.DataFrame(columns = ['cd_ibge','vl_emissao'])
df_emis['cd_ibge'] = predictions_emissoes.keys()
df_emis['vl_emissao'] = predictions_emissoes.values()

Juntando os 2 dataframes

predicoes = pd.merge(df_arrec, df_emis, how='inner', on='cd_ibge')

Adicionando informações de estado e cidade

predicoes = pd.merge(predicoes, municipios, how='inner', on='cd_ibge')

predicoes['dif'] = predicoes['vl_arrecadado'] - predicoes['vl_emissao']

predicoes['indice'] = predicoes['vl_arrecadado'] / predicoes['vl_emissao']

predicoes

df.loc[df['cd_ibge']==3506201]
